v2

+++ lib/home.dart
@@ -1,10 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:kmel_side_app/fire_s.dart';
-
-// Map<String,bool> selectedMap = Map<String, bool>();
-// List<bool> selected = [];
 
 
 class HomePage extends StatefulWidget {
@@ -18,11 +14,20 @@ class _HomePageState extends State<HomePage> {
 
   deletePrevApps(BuildContext context) async {
     Navigator.pushNamed(context, '/load');
-    DateTime yDay = DateTime.now().subtract(Duration(days: 1));
-    QuerySnapshot yesterdayDocs = await FirebaseFirestore.instance.collection('appointments').where('date',isEqualTo: '${yDay.day}/${yDay.month}/${yDay.year}').get();
-    List<DocumentSnapshot> docs = yesterdayDocs.docs;
-    for(int i=0 ; i < docs.length ; i++){
-      await db.deleteApp(docs[i].id);
+    DateTime today = DateTime.now();
+    for (int j=1 ; j < 1000 ; j++) {
+      DateTime dayToDelete = today.subtract(Duration(days: j));
+      if (dayToDelete.weekday == DateTime.monday) {
+        continue;
+      }
+      QuerySnapshot dayToDeleteDocs = await FirebaseFirestore.instance.collection('appointments').where('date',isEqualTo: '${dayToDelete.day}/${dayToDelete.month}/${dayToDelete.year}').get();
+      List<DocumentSnapshot> docs = dayToDeleteDocs.docs;
+      if (docs.length == 0) {
+        break;
+      }
+      for(int i=0 ; i < docs.length ; i++){
+        await db.deleteApp(docs[i].id);
+      }
     }
     Navigator.of(context).pop();
   }
@@ -32,10 +37,11 @@ class _HomePageState extends State<HomePage> {
   Widget build(BuildContext context) {
     return SafeArea(
       child: Scaffold(
-        resizeToAvoidBottomPadding: false,
+        resizeToAvoidBottomInset: false,
         backgroundColor: Colors.grey[800],
         body: Container(
           width: MediaQuery.of(context).size.width * 0.98,
+
           height: MediaQuery.of(context).size.height,
           child: Center(
             child: Column(
@@ -55,25 +61,13 @@ class _HomePageState extends State<HomePage> {
                       padding: const EdgeInsets.all(8.0),
                       child: FlatButton(
                         onPressed: () async {
-
                           await deletePrevApps(context);
-
-
-                            // int docs_num_today = await db.initSelectedNum();
-                            // selected=List<bool>();
-                            // for(int i=0 ; i < docs_num_today ; i++){
-                            //   selected.add(false);
-                            //   print(i);
-                            // }
-
-
-
-                          Navigator.pushNamed(context, '/view');
+                          Navigator.pushNamed(context, '/preViewDate');
                         }, // change this to real function
                         child: Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: Text(
-                            'כניסה למערכת',
+                            'כניסה למערכת תורים',
                             style: TextStyle(
                               fontSize: 20,
                               color: Colors.white,
@@ -91,10 +85,33 @@ class _HomePageState extends State<HomePage> {
                   ],
                 ),
 
-
-
-
-
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: FlatButton(
+                        onPressed: () async {
+                          Navigator.pushNamed(context, '/currentVacations');
+                        }, // change this to real function
+                        child: Padding(
+                          padding: const EdgeInsets.all(8.0),
+                          child: Text(
+                            'כניסה למערכת חופשים',
+                            style: TextStyle(
+                              fontSize: 20,
+                              color: Colors.white,
+                            ),
+                          ),
+                        ),
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(18),
+                          side: BorderSide(color: Colors.grey),
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
 
               ],
             ),
